+++ test/astar.cc
@@ -1632,6 +1632,60 @@ TEST(ComplexRestriction, WalkVias) {
   }
 }
 
+TEST(Astar, BiDirTrivial) {
+  // Normally the service does not allow a trivial path with bidirectional astar because it has some
+  // problems with those (oneways that make you go around the block to get to where you started?).
+  // However when reviewing the other special case of short bidirectional routes where the forward and
+  // backward search paths meet on the destination edge we found that the trivial path edge trimming
+  // was wrong. Specifically because the forward expansion only cares about trimming the first edge
+  // from the origin and the reverse expansion only cares about trimming the last edge up to the
+  // destination but the route is only one edge. This means that both the reverse path label and the
+  // forward path label both have trimmed the edge but not enough. So what we have to do is trim the
+  // whole edge based on what percentage of the edge is left between the origin and destination.
+
+  // Get access to tiles
+  boost::property_tree::ptree conf;
+  conf.put("tile_dir", "test/data/utrecht_tiles");
+  vb::GraphReader graph_reader(conf);
+
+  // Locations
+  std::vector<valhalla::baldr::Location> locations;
+  baldr::Location origin(valhalla::midgard::PointLL(5.12696, 52.09701),
+                         baldr::Location::StopType::BREAK);
+  locations.push_back(origin);
+  baldr::Location dest(valhalla::midgard::PointLL(5.12700, 52.09709),
+                       baldr::Location::StopType::BREAK);
+  locations.push_back(dest);
+
+  // Costing
+  valhalla::Options options;
+  create_costing_options(options);
+  std::shared_ptr<vs::DynamicCost> mode_costing[4];
+  std::shared_ptr<vs::DynamicCost> cost = vs::CreateAutoCost(Costing::auto_, options);
+  auto mode = cost->travel_mode();
+  mode_costing[static_cast<uint32_t>(mode)] = cost;
+
+  // Loki
+  const auto projections = vk::Search(locations, graph_reader, cost);
+  std::vector<PathLocation> path_location;
+  for (const auto& loc : locations) {
+    ASSERT_NO_THROW(
+        path_location.push_back(projections.at(loc));
+        PathLocation::toPBF(path_location.back(), options.mutable_locations()->Add(), graph_reader);)
+        << "fail_invalid_origin";
+  }
+
+  vt::BidirectionalAStar astar;
+  auto path = astar
+                  .GetBestPath(*options.mutable_locations(0), *options.mutable_locations(1),
+                               graph_reader, mode_costing, mode)
+                  .front();
+
+  ASSERT_TRUE(path.size() == 1);
+  EXPECT_LT(path.front().elapsed_cost, 1);
+  EXPECT_LT(path.front().elapsed_time, 1);
+}
+
 class AstarTestEnv : public ::testing::Environment {
 public:
   void SetUp() override {